refactor: rename JsonResponse to ResponseCreatorTestTool and move to tests
- Remove src/Response/JsonResponse.php (no longer part of public API)
- Create tests/Tools/ResponseCreatorTestTool with mandatory $status param
- Rename method from from() to createJsonResponse()
- Update all test sample usages
