+++ tests/Sample/CreateChannelRoute.php
@@ -3,7 +3,7 @@
 namespace BrandEmbassyTest\Slim\Sample;
 
 use BrandEmbassy\Slim\Request\RequestInterface;
-use BrandEmbassy\Slim\Response\JsonResponse;
+use BrandEmbassyTest\Slim\Tools\ResponseCreatorTestTool;
 use BrandEmbassy\Slim\Response\ResponseInterface;
 use BrandEmbassy\Slim\Route\Route;
 
@@ -14,6 +14,6 @@ class CreateChannelRoute implements Route
 {
     public function __invoke(RequestInterface $request, ResponseInterface $response): ResponseInterface
     {
-        return JsonResponse::from($response, ['status' => 'created'], 201);
+        return ResponseCreatorTestTool::createJsonResponse($response, ['status' => 'created'], 201);
     }
 }
